+++ apps/src/code-studio/components/libraries/LibraryCreationDialog.jsx
@@ -16,12 +16,12 @@ import i18n from '@cdo/locale';
 import Spinner from '../../../sharedComponents/Spinner';
 import {hideLibraryCreationDialog} from '../shareDialogRedux';
 
+import {extractTextFromCode} from './extractTextFromCode';
 import LibraryClientApi from './LibraryClientApi';
 import loadLibrary from './libraryLoader';
 import LibraryPublisher from './LibraryPublisher';
 import PublishSuccessDisplay from './PublishSuccessDisplay';
 import ShareTeacherLibraries from './ShareTeacherLibraries';
-import {extractTextFromCode} from './utils';
 
 const DEFAULT_COPY_BUTTON_TEXT = i18n.copyId();
