+++ apps/src/code-studio/components/libraries/extractTextFromCode.ts
